+++ crates/browser-ui/src/events/kernel/response/devtools.rs
@@ -30,8 +30,8 @@ pub fn on_devtools_page_ready(application: &mut Application, window_id: Id, tab_
         let localhost = Url::parse(&format!("http://{}", Ipv4Addr::LOCALHOST)).unwrap();
         let abs_ctx = AbsoluteContext {
             root_font_size: 16.0,
-            viewport_width: devtools_ctx.viewport.width,
-            viewport_height: devtools_ctx.viewport.height,
+            viewport_width: f64::from(devtools_ctx.viewport.width),
+            viewport_height: f64::from(devtools_ctx.viewport.height),
             theme_category: application.config.preferences().theme().category,
             document_url: &localhost,
             root_line_height_multiplier: 1.2,
@@ -44,7 +44,7 @@ pub fn on_devtools_page_ready(application: &mut Application, window_id: Id, tab_
         let layout_tree = LayoutEngine::compute_layout(
             page.document(),
             &style_tree,
-            Rect::new(0.0, 0.0, devtools_ctx.viewport.width, devtools_ctx.viewport.height),
+            Rect::new(0.0, 0.0, f64::from(devtools_ctx.viewport.width), f64::from(devtools_ctx.viewport.height)),
             &mut tc,
             &img_ctx,
         );
